Workflow to update downstream dataset repos
Tentative workflow to make a PR to the main branch of dataset-specific repos. Triggered whenever a commit is pushed to the `main` branch of the main `mr_proc` repo (since the branch is protected, this means that it will trigger when PRs are merged to `main`.

Need to test this on a toy repo.